services/collector-yfinance/requirements.txt
yfinance>=0.2.48
pika==1.3.2
python-dotenv==1.0.1
pandas>=1.5.0
requests>=2.28.0
scipy>=1.9.0
